Add ids in some classes and json examples

+++ include/Point.h
@@ -9,13 +9,13 @@ class Point
     public:
 
         Point();
-        Point(short pos_x, short pos_y);
+        Point(int pos_x, int pos_y);
 
         static Point GET_ORIGIN() { return ORIGIN; }
         static void SET_ORIGIN(Point val) { ORIGIN = val; }
 
-        short pos_x() const { return m_pos_x; }
-        short pos_y() const { return m_pos_y; }
+        int pos_x() const { return m_pos_x; }
+        int pos_y() const { return m_pos_y; }
 
         std::string to_string() const;
 
@@ -24,14 +24,19 @@ class Point
 
         static Point ORIGIN;
 
-        short m_pos_x; // in m
-        short m_pos_y; // in m
+        int m_pos_x; // in m
+        int m_pos_y; // in m
 };
 
 bool operator==(Point const& point1, Point const& point2);
 std::ostream& operator<<(std::ostream &flux, Point const& point);
 
-// Manhattan distance
-long distance(Point const& point1, Point const& point2);
+/**
+ * Calculate the Manhattan distance between two points
+ * @param point1 The first point
+ * @param point2 The second point
+ * @return The distance between the two points
+ */
+int distance(Point const& point1, Point const& point2);
 
 #endif // POINT_H
